Renamed BodyParser to StreamingBodyParser to make it more explicit what those parsers are doing
