+++ src/StreamingBodyParser/StreamingParserInterface.php
@@ -1,11 +1,11 @@
 <?php
 
-namespace React\Http\BodyParser;
+namespace React\Http\StreamingBodyParser;
 
 use Evenement\EventEmitterInterface;
 use React\Http\Request;
 
-interface ParserInterface extends EventEmitterInterface
+interface StreamingParserInterface extends EventEmitterInterface
 {
     public function __construct(Request $request);
 }
